fix: handle int64 Bavail type on FreeBSD
On FreeBSD, stat.Bavail is int64 while on Linux it's uint64.
Cast both operands to uint64 for cross-platform compatibility.

+++ cmd/bd/daemon_health_unix.go
@@ -15,8 +15,9 @@ func checkDiskSpace(path string) (uint64, bool) {
 	}
 
 	// Calculate available space in bytes, then convert to MB
-	// Bavail is uint64, Bsize is int64; overflow is intentional/safe in this context
-	availableBytes := stat.Bavail * uint64(stat.Bsize) //nolint:gosec
+	// Bavail type varies by OS (uint64 on Linux, int64 on FreeBSD/Darwin)
+	// Bsize is int64; overflow is intentional/safe in this context
+	availableBytes := uint64(stat.Bavail) * uint64(stat.Bsize) //nolint:gosec
 	availableMB := availableBytes / (1024 * 1024)
 
 	return availableMB, true
